requirements.txt
discord.py
aiohttp
python-dotenv
requests
py-cord==2.6.1
tweepy
beautifulsoup4
lxml
psycopg2-binary
matplotlib
pillow
